Move AmbiguousFlagTrustee and detectAmbiguousFlagTrustees to adapter layer
Both operate entirely on raw ATS SQL column names and belong in the
adapter/cleansing layer alongside transformTrusteeRecord. Moves them
from migrate-trustees.ts to ats-mappings.ts and updates all imports.
Removes redundant dispOnWeb/dispOnWebA2 fields from AmbiguousFlagTrustee
— condition already encodes the same information.
